requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
click==8.1.8
fastapi==0.115.8
greenlet==3.1.1
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
protobuf==5.29.3
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
python-dotenv==1.0.1
setuptools==75.8.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
